+++ README.md
@@ -2,12 +2,12 @@
 
 Benchmarking tools for LanceDB Cloud and LanceDB Enterprise.
 
-### Background 
+### Background
 
 This benchmark script will download the [dbpedia-entities-openai-1M](https://huggingface.co/datasets/KShivendu/dbpedia-entities-openai-1M) dataset,
 ingest and index it into N tables in LanceDB Cloud/Enterprise and run vector searches on the tables.
 
-It will report on ingestion time, indexing completion time, and query performance percentiles. 
+It will report on ingestion time, indexing completion time, and query performance percentiles.
 
 Further metrics can be gathered by the [LanceDB team](mailto:contact@lancedb.com) upon request.
 
@@ -33,7 +33,7 @@ export LANCEDB_HOST_OVERRIDE=<optional uri if using lancedb enterprise>`
 
 ### Examples
 
-Ingest the dataset into 4 tables and run 10k queries with a custom table prefix: 
+Ingest the dataset into 4 tables and run 10k queries with a custom table prefix:
 
 `uv run bench.py -t 4 -q 10000 -p mytable`
 
@@ -57,19 +57,48 @@ At high traffic levels, ingestion and query performance may be limited in a sing
 to larger aggregate numbers by using multiple processes or even distributing across multiple VMs. In this case, the result metrics will need to be aggregated
 to get the total QPS and throughput.
 
-Example:
-To run multiple query benchmarks in parallel processes, here’s an example command that initiates 4 benchmarking processes,
-each querying 4 tables, using the table name prefix "my-prefix":
+#### Example 1: Each Process Creates a Table and Queries It
+To run multiple query benchmarks in parallel processes, you can initiate several benchmarking processes,
+each creating and querying its own tables.
+The following command demonstrates how to start 4 benchmarking processes,
+each querying 4 tables with the table name prefix "my-prefix":
 ```
-uv run multi.py -n 4 -t 4 -p my-prefix -q 10000 -r
+uv run bench.py -n 4 -t 4 -p my-prefix -q 10000 -r
 ```
+Parameters:
+`-n 4`: Number of benchmarking processes to run
+`-t 4`: Number of tables to query per process
+`-p my-prefix`: Prefix for the table names
+`-q 10000`: Number of queries to run against each table
+`-r`: recreate the table and indices if exist
 
-After the initial setup, you can rerun the query performance tests without recreating tables or indexes by using:
+After the initial setup, you can rerun the query performance tests without recreating tables or indexes by using the following command:
 ```
-uv run multi.py -n 4 -t 4 -p my-prefix -q 10000 --no-ingest --no-index
+uv run bench.py -n 4 -t 4 -p my-prefix -q 10000 --no-ingest --no-index
 ```
+Additional Flags:
+`--no-ingest`: Skips the table creation step.
+`--no-index`: Skips the index creation step.
 
-For additional options and usage details, run:
+#### Example 2: Multiple Processes Querying the Same Table
+In this scenario, you first create a table with a specified name prefix and the necessary indexes.
+Use the following command to set this up:
 ```
-uv run multi.py --help
+uv run bench.py -t 1 -p my-prefix -q 0 -r
 ```
+Parameters:
+`-t 1`: Create one table (the same table will be queried by multiple processes).
+`-p my-prefix`: Prefix for the table name.
+`-q 0`: No queries are run during the table creation.
+`-r`: recreate the table and indices if exist
+
+Once the table is created, you can launch multiple processes to query against the same table.
+Each process will run 10,000 queries. Use the following command:
+```
+uv run bench.py -t 1 -p my-prefix --no-ingest --no-index --query-process 5 -q 10000
+```
+Parameters:
+`--query-process 5`: Specifies that 5 processes will query the same table concurrently.
+`--no-ingest`: Skips the table creation step.
+`--no-index`: Skips the index creation step.
+`-q 10000`: Number of queries to run against each table
