+++ bench.py
@@ -1,9 +1,12 @@
 import argparse
 import concurrent
 import os
+import sys
 import time
 from concurrent.futures import wait
-from typing import Iterable, List
+import traceback
+from typing import Iterable, List, Tuple
+import multiprocessing as mp
 
 from lancedb.remote.errors import LanceDBClientError
 from lancedb.remote.table import RemoteTable
@@ -122,8 +125,8 @@ class Benchmark:
         if self.index:
             self._create_indices()
 
-        # TODO(lu) add self.query
-        self._query_tables()
+        if self.num_queries > 0:
+            self._query_tables()
         return self.results
 
     def _create_tables(self) -> Iterable[RemoteTable]:
@@ -375,7 +378,30 @@ class Benchmark:
             self.results.ingest_latencies.extend(diffs)
 
 
-def run_benchmark(
+def run_benchmark_process(process_args: Tuple[int, int, dict]) -> str:
+    """Run a single benchmark process
+    Args:
+        process_args: Tuple of (process_id, query_id, results, bench_kwargs)
+    """
+    process_id, query_id, bench_kwargs = process_args
+    try:
+        # Modify prefix for this process group
+        bench_kwargs = bench_kwargs.copy()
+        bench_kwargs["prefix"] = f"{bench_kwargs['prefix']}-{process_id}"
+
+        benchmark = Benchmark(**bench_kwargs)
+        result = benchmark.run()
+        return result.to_json()
+
+    except Exception as e:
+        print(f"Process {process_id}, query {query_id} failed: {e}", file=sys.stderr)
+        print(traceback.format_exc(), file=sys.stderr)
+        return None
+
+
+def run_multi_benchmark(
+    num_processes: int,
+    query_process: int,
     dataset: str,
     num_tables: int,
     batch_size: int,
@@ -385,19 +411,83 @@ def run_benchmark(
     prefix: str,
     reset: bool,
 ) -> BenchmarkResults:
-    benchmark = Benchmark(
-        dataset, num_tables, batch_size, num_queries, ingest, index, prefix, reset
-    )
-    return benchmark.run()
+    total_processes = num_processes * (query_process if not ingest and not index else 1)
+    print(f"Starting {total_processes} benchmark processes...")
+
+    bench_kwargs = {
+        "dataset": dataset,
+        "num_tables": num_tables,
+        "batch_size": batch_size,
+        "num_queries": num_queries,
+        "ingest": ingest,
+        "index": index,
+        "prefix": prefix,  # Base prefix, will be modified per process
+        "reset": reset,
+    }
+
+    process_args = []
+
+    if ingest or index:
+        for i in range(0, num_processes):
+            process_kwargs = bench_kwargs.copy()
+            process_args.append((i, 0, process_kwargs))
+    else:
+        for i in range(0, num_processes):
+            for j in range(0, query_process):
+                process_kwargs = bench_kwargs.copy()
+                process_args.append((i, j, process_kwargs))
+
+    with mp.Pool(processes=total_processes) as pool:
+        process_results = pool.map(run_benchmark_process, process_args)
+
+        successful_results = [
+            BenchmarkResults.from_json(r) for r in process_results if r is not None
+        ]
+
+        if not successful_results:
+            raise RuntimeError(
+                "All benchmark processes failed - check logs for details"
+            )
+
+        return BenchmarkResults.combine(successful_results)
+
+
+def validate_args(args: argparse.Namespace):
+    if args.query_process > 1:
+        if args.ingest or args.index:
+            raise ValueError(
+                "Multiple query processes per table (query_process > 1) is only allowed "
+                "with --no-ingest and --no-index flags"
+            )
 
 
 def main():
     parser = argparse.ArgumentParser()
+    parser.add_argument
+    parser.add_argument(
+        "-n",
+        "--num-processes",
+        type=int,
+        required=False,
+        default=1,
+        help="Number of total benchmark process. This number should be the same for data ingestion and data querying.",
+    )
+    parser.add_argument(
+        "-qn",
+        "--query-process",
+        type=int,
+        required=False,
+        default=1,
+        help="Number of the query process per table. When this number is not 1, total process number becomes num_processes * query_process.",
+    )
     add_benchmark_args(parser)
     args = parser.parse_args()
+    validate_args(args)
     print(args)
 
-    result: BenchmarkResults = run_benchmark(
+    result = run_multi_benchmark(
+        args.num_processes,
+        args.query_process,
         args.dataset,
         args.tables,
         args.batch,
@@ -407,7 +497,8 @@ def main():
         args.prefix,
         args.reset,
     )
-    result.print(True)
+
+    result.print()
 
 
 if __name__ == "__main__":
